requirements.txt
amqp==5.1.1
asgiref==3.4.1
async-timeout==4.0.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.13
-e git+https://github.com/celery/django-celery-results.git@ed94080cea3d68a38f9ff048657a37b08e74fc35#egg=django_celery_results
djangorestframework==3.13.1
importlib-metadata==4.8.3
kombu==5.1.0
packaging==21.3
prompt-toolkit==3.0.29
pyparsing==3.0.9
pytz==2022.1
redis==4.3.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.1.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.6.0
